lambda_function/requirements.txt
databricks-sql-connector>=4.0.0
httpx>=0.27.0
